Modified coordinate frames and added Euler angle stuff
Euler angles are currently not investigated as far as how good they are coming out - we need to have commands put in first and real dynamics before we can do anything. cleaned up the code so that it's easier and more user-readable. also included some files I'd forgotten to add before.